+++ src/test/kotlin/com/zepben/ewb/services/network/NetworkServiceComparatorOptionsTest.kt
@@ -24,23 +24,23 @@ internal class NetworkServiceComparatorOptionsTest {
     @Test
     internal fun all() {
         val options = NetworkServiceComparatorOptions.all()
-        assertThat("compareTracedPhases should be true for all()", options.compareTracedPhases)
-        assertThat("compareTerminals should be true for all()", options.compareTerminals)
-        assertThat("compareFeederEquipment should be true for all()", options.compareFeederEquipment)
-        assertThat("compareEquipmentContainers should be true for all()", options.compareEquipmentContainers)
-        assertThat("compareLvSimplification should be true for all()", options.compareLvSimplification)
-        assertThat("compareRunTimeVariables should be true for all()", options.compareRunTimeVariables)
+        assertThat("compareTracedPhases should be enabled for all()", options.compareTracedPhases)
+        assertThat("compareTerminals should be enabled for all()", options.compareTerminals)
+        assertThat("compareFeederEquipment should be enabled for all()", options.compareFeederEquipment)
+        assertThat("compareEquipmentContainers should be enabled for all()", options.compareEquipmentContainers)
+        assertThat("compareLvSimplification should be enabled for all()", options.compareLvSimplification)
+        assertThat("compareRunTime should be enabled for all()", options.compareRunTime)
     }
 
     @Test
     internal fun none() {
         val options = NetworkServiceComparatorOptions.none()
-        assertThat("compareTracedPhases should be false for none()", !options.compareTracedPhases)
-        assertThat("compareTerminals should be false for none()", !options.compareTerminals)
-        assertThat("compareFeederEquipment should be false for none()", !options.compareFeederEquipment)
-        assertThat("compareEquipmentContainers should be false for none()", !options.compareEquipmentContainers)
-        assertThat("compareLvSimplification should be false for none()", !options.compareLvSimplification)
-        assertThat("compareRunTimeVariables should be false for none()", !options.compareRunTimeVariables)
+        assertThat("compareTracedPhases should be disabled for none()", !options.compareTracedPhases)
+        assertThat("compareTerminals should be disabled for none()", !options.compareTerminals)
+        assertThat("compareFeederEquipment should be disabled for none()", !options.compareFeederEquipment)
+        assertThat("compareEquipmentContainers should be disabled for none()", !options.compareEquipmentContainers)
+        assertThat("compareLvSimplification should be disabled for none()", !options.compareLvSimplification)
+        assertThat("compareRunTime should be disabled for none()", !options.compareRunTime)
     }
 
     @Test
@@ -51,26 +51,36 @@ internal class NetworkServiceComparatorOptionsTest {
             .compareFeederEquipment()
             .compareEquipmentContainers()
             .compareLvSimplification()
-            .compareRunTimeVariables()
             .build()
 
-        assertThat("compareTracedPhases should be true when building with all options enabled", options.compareTracedPhases)
-        assertThat("compareTerminals should be true when building with all options enabled", options.compareTerminals)
-        assertThat("compareFeederEquipment should be true when building with all options enabled", options.compareFeederEquipment)
-        assertThat("compareEquipmentContainers should be true when building with all options enabled", options.compareEquipmentContainers)
-        assertThat("compareLvSimplification should be true when building with all options enabled", options.compareLvSimplification)
-        assertThat("compareRunTimeVariables should be true when building with all options enabled", options.compareRunTimeVariables)
+        assertThat("compareTracedPhases should be enabled when building with all options enabled", options.compareTracedPhases)
+        assertThat("compareTerminals should be enabled when building with all options enabled", options.compareTerminals)
+        assertThat("compareFeederEquipment should be enabled when building with all options enabled", options.compareFeederEquipment)
+        assertThat("compareEquipmentContainers should be enabled when building with all options enabled", options.compareEquipmentContainers)
+        assertThat("compareLvSimplification should be enabled when building with all options enabled", options.compareLvSimplification)
+
+        // You can't turn on the new `compareRunTime` option via the deprecated builder.
+        assertThat("compareRunTime should be disabled in the deprecated builder", !options.compareRunTime)
     }
 
     @Test
     internal fun builderNone() {
         val options = NetworkServiceComparatorOptions.of().build()
 
-        assertThat("compareTracedPhases should be false when building without any options enabled", !options.compareTracedPhases)
-        assertThat("compareTerminals should be false when building without any options enabled", !options.compareTerminals)
-        assertThat("compareFeederEquipment should be false when building without any options enabled", !options.compareFeederEquipment)
-        assertThat("compareEquipmentContainers should be false when building without any options enabled", !options.compareEquipmentContainers)
-        assertThat("compareLvSimplification should be false when building without any options enabled", !options.compareLvSimplification)
-        assertThat("compareRunTimeVariables should be false when building without any options enabled", !options.compareRunTimeVariables)
+        assertThat("compareTracedPhases should be disabled when building without any options enabled", !options.compareTracedPhases)
+        assertThat("compareTerminals should be disabled when building without any options enabled", !options.compareTerminals)
+        assertThat("compareFeederEquipment should be disabled when building without any options enabled", !options.compareFeederEquipment)
+        assertThat("compareEquipmentContainers should be disabled when building without any options enabled", !options.compareEquipmentContainers)
+        assertThat("compareLvSimplification should be disabled when building without any options enabled", !options.compareLvSimplification)
+        assertThat("compareRunTime should be disabled in the deprecated builder", !options.compareRunTime)
     }
+
+    @Test
+    internal fun constructorCoverage() {
+        // Stupid test to give coverage to generated Kotlin code that is guaranteed to work.
+        NetworkServiceComparatorOptions()
+        NetworkServiceComparatorOptions(compareTerminals = false)
+        NetworkServiceComparatorOptions(compareRunTime = false)
+    }
+
 }